Ship Selection
Nearing completion of the ship selection aspect of new games.

+++ Scavenger/Assets/Ships/Scripts/ShipDefinition.cs
@@ -18,6 +18,6 @@ public struct ShipDefinition
     public string material;
     public string risk;
     public int oneIn;
-    public float value;
+    public int value;
     public List<Statistic> statistics;
 }
